Bug 346 - Port to PPC/Linux
This provides PowerPC/Linux support for Pegasus with the target name
LINUX_PPC_GNU.

Much of this code is generic, and indeed a lot of my changes in
the files were to convert sections that were
  #ifdef LINUX_IX86_GNU
to instead
  #if defined(LINUX_IX86_GNU)  || defined (LINUX_GENERIC_GNU)
It strikes me that much more clean up and generalization of this
order could be done among the Linux ports.

NOTE: this port uncovered a bug in glibc on PPC/Linux -- the fix will be in the
next release of glibc.	You can get the fix temporarily from glibc CVS (see
http://sources.redhat.com/glibc/) or from me at dteger@us.ibm.com