BUG#: 1235 BRANCH: RELEASE_2_3_2-branch TITLE: providerDir only supports one directory. Should support more directories.
DESCRIPTION:
Microsoft Windows has in interesting loading mechanism. It loads the library file from the current directory, and if that fails - it searches in the specified environment variables. But this fix does not depend on the loading mechanism, instead it checks to see if the file exist and then passes the complete path to the provider loader. Therefore to make Windows load the providers, the directory 'bin' (where the dll's for providers are located) has to be appended to the default 'providerDir' parameter.