BUG#: 1319 TITLE: SSL client authentication not fully implemented
DESCRIPTION

--Removing a slight change to the server verification function in
  SSLSocket's connect() call in order to decrease risk for SLES9 beta.
  Per today's meeting, the following issue can be addressed in a new 2.3.2 PEP.
  Note that the implementation for this is still wrong, but I'm setting it
  back so it behaves as before.

Using SSL_get_verify_result is not a valid way to determine whether or not to
continue the handshake because the result is equal to the preverify error,
and not the value after the SSL callback function.

The verification function that the client can use in its SSLContext should
be used instead.