BUG#: 1354 TITLE: HP [CAKG] IndicationService should use SendAsync to send handle indication request to Handler
DESCRIPTION: The IndicationService should use SendAsync (instead of SendWait) to send the Handle Indication request message to the Handler.  The SendWait is inside a for loop that is sending a request for each matching subscription.  If there are many subscriptions for an indication, using SendWait could slow down indication processing performance.  The subscription's OnFatalErrorPolicy data is needed when the response is received, so a userParameter (the subscription instance) is included in the callback.