BUG#: 1435 TITLE: [HP-TM] Order of Include Files Affects Windows Build
DESCRIPTION: Windows can be very sensitive to the order in which certain system files are included, particularly between Windows.h and Winsock.h.  When using the Microsoft Platform SDK, this is even more of a problem, so that certain files will not build without changing the include orderings on these files.