BUG#: 1400 TITLE: Check if HTTPS is enabled before creating SSLContext for indications
DESCRIPTION:

Previously the SSLContext object was being created for indication messages whether or not HTTPS was being used, or even enabled. This resulted in errors when the CIMOM tried to look up the server's SSL certificate and one is not availalbe. If HTTPS is enabled, then this condition is checked at startup. The SSLContext will only be created now if HTTPS is enabled and the destination is an HTTPS location.