PEP#: 072 TITLE: [HP-RK] Out-of-Process Provider Support
DESCRIPTION:

Move some of the functionality of the ProviderManagerService into a new ProviderManagerRouter module.  The ProviderManagerRouter is an abstract class, with a BasicProviderManagerRouter subclass implementing in-process provider support.  Out-of-Process Provider functionality will be encapsulated in a separate OOPProviderManagerRouter class.

This change also adds a callback function for indication delivery, rather than having the EnableIndicationsResponseHandler interact directly with the MessageQueueService.  The callback approach more clearly delineates the functions of the ProviderManagerService and the ProviderManagers, and facilitates the generation of indications by Out-of-Process Providers.

ProviderManagers are now loaded on demand, rather than automatically at start-up time.  This change will reduce the footprint of Provider Agent Processes.