PEP#: 072 TITLE: [HP-RK] Out-of-Process Provider Support
DESCRIPTION:

Added a CIMOMHandle implementation for use by Out-of-Process Providers.  This implementation uses a CIMClient object to forward CIM operation requests to the CIM Server.

The CIMOMHandle implementation was reorganized with this change.  An abstract CIMOMHandleRep class allows for polymorphic implementations.  The InternalCIMOMHandleRep encapsulates the existing, in-process CIMOMHandle implementation.  A ClientCIMOMHandleRep class contains the new, CIMClient-based implementation.