PEP#: 155 TITLE: HP[Nag] SSL client certificate verification for CIMExport requests
DESCRIPTION: Added SSL client certificate verification support for CIMExport requests received on the new export port. To enable this feature set the config property enableSSLExportClientVerification to true and specify the indication trust store using the config property exportSSLTrustStore. Username password authentication is not done and CIMOperation requests are not processed on the export port.