PEP#:072 TITLE: [HP-YZ] Out-of-Process Provider Support
DESCRIPTION: Added a new property AutoStart in the PG_Provider class and new functions have been added to support this new property. At CIM Server start-up, if AutoStart property is set to true in a PG_Provider instance, the Pluggable Provider Manager will load and intialize the provider. This feature applies to existing In-Process Providers as well as Out-of-Process Providers.