PEP#: 142 TITLE: HP[Nag] User Groups based authorization
DESCRIPTION: Added user groups based authorization support to Pegasus. This is an additional authorization mechanism that gives an administrator a simple, overriding control over which users may perform CIM operations. To enable this feature, define the compile time flag PEGASUS_ENABLE_USERGROUP_AUTHORIZATION and set the config property authorizedUserGroups to one or more valid user groups on the system. (This feature is not fully implemented on Windows, just a skeleton is defined for isGroupMember() function in systemWindows.cpp file.)