PEP#: 072 TITLE: [HP-RK] Out-of-Process Provider Support
DESCRIPTION: Remove CIMRepository dependency from the CMPI and JMPI Provider Managers.  (The Provider Managers will be used in Provider Agent processes, which do not have access to the repository.)  The repository had been required only to retrieve the query string from the CIM_IndicationFilter instance on a CreateSubscription operation.  The query string has been added to the CIMCreateSubscriptionRequestMessage and CIMModifySubscriptionRequestMessage classes to meet this need.