BUG#: 1496 TITLE: CBDeliverIndicaction
DESCRIPTION:
The CMPI_ThreadContext::getBroker() does a 'pegasus_get_thread_specific(key)'
to get the CMPI_ThreadContext object and return it back and dereference to the
broker. However, the result of the 'pegasus_get_thread_specific(key)' is not a
CMPI_ThreadContext object, but 0, so the derefrencing causes a seg fault.

There is a set of four functions that utilize this: CBPrepareAttachThread, CBAttachThread, mbDeliverIndiacation,CBDetachThread. All of the tried to obtain the Broker from the TSD, which is incorrect. The CBAttachThread should _not_ try to obtain the Broker from the TSD. The previous fixes uncommented the wrong line, this fix uncomments the right line and fixes the previous fix.