PEP#: 072 TITLE: [HP-RK] Out-of-Process Provider Support
DESCRIPTION: Updated the ConfigManager so that it can be used in a Provider Agent process (for Out-of-Process Providers).  The ConfigManager now has a flag that controls whether it reads to and writes from persistent configuration files (for the cimserver process) or it just operates in memory (for Provider Agent processes and other uses).