BUG#: 1686 TITLE: Chunking doesnt seem to handle server errors correctly DESCRIPTION:
problem:
a. The full error description was not coming back to client.
b. Content language was not being communicated to client on an error condition.

solution:
a. call TraceableCIMException().getDescription() to expand the error text.
   When chunking was implemented, there was a call that originally placed the
   error in the XML payload, but the cim operations spec dictates that the XML
   payload is to be ignored on error conditions, however, the full error still
   needs to be placed into the CIMStatusCodeDescription field in the HTTP
   trailer to communicate the error outside of the XML payload.
b. call setContentLanguages() on the cim exception object

files fixed:
a. Server/CIMOperationResponseEncoder.cpp
b. Client/CIMOperationResponseDecoder.cpp