BUG#: 1691 TITLE: ObjectNormalizer bugs
DESCRIPTION: Changed dispatcher code to aggregate responses in first response (as orginally coded),
and then normalize after. Normalized instances replace the unnormalized instances in the first
response, and objects that fail to normalize are dropped (error is logged). Eventaully, providers
should be notified when an object fails to normalize, but the current aggregation comes much to late.