PEP#: 1732 TITLE:  Pegasus implements SLP security and does not allow the security to be configured dynamically
DESCRIPTION:  Fixed this for version 2.4 by blocking all security.  This patch was provided by M. Day. We noted that because of changes that are occuring in security concepts in SLP it was probably better to block rather than do a dynamic fix unless someone specifically wanted DA security implemented.  We will leave this as a permanent fix unless there is a real need for DA security.