BUG#: 1767 TITLE: Problem with AutoPtr and no destructor.
DESCRIPTION: Fixed the problem here by moving the type declarations from the source file into the header file and removing the forward declaration.   The type declarations are private to the class so there is no namespace pollution.  With the types fully declared they now have destructors and the warning is no longer issued.