BUG#: 1765 TITLE: CLI instance output ignores properties
DESCRIPTION: Fixed the CLI to output instances in MOF format as it does classes. That is,
set all properties to PROPAGATED = false so the MofWriter shows them. This causes the
CLI to show the entire instance/class returned, rather than just the leaf properties needed
in the generatation of an actual MOF.