BUG#: 1740
TITLE:  Update methods to reflect Template.

DESCRIPTION:  Updated the methods to reflect the PG_wbemslpTemplate class change.  Note that we also fixed several other CRs in the
same update including:  1585, case on class info, 1547, profile component separators.  Also, reorganized the set of constants to be more consistent
with the template and to simplify the generation of each attribute.