BUG#: 1849 TITLE:  Modify make files so that PEGASUS_DEBUG becomes a common compiler DEFINE available through config.mak
DESCRIPTION:  This involved removing the setting of PEGASUS_DEBUG into a number of platform mak setup files and putting the DEFINE into config.mak.  This means that PEGASUS_DEBUG now generally controls a) setting the debug options on compiles and links in general as it did before and is also available to any source code to compile diagnostics.  However, since this is very general we are trying to reduce the use to those debug concepts that will not affect normal operation.  Thus it is used to control the output of startup information to the console.  PLEASE do not use it to control things like debug information output to the console.