BUG#: 1820 TITLE: EnumInstances test fails
DESCRIPTION: Updated ObjectNormalizer to *try* to complete the object path
if the provider does not specify it. This is not always possible because an operation
may request a partial object where some or all keys are missing. In this case, the default
values (as specified by the type, not class definition) are used. For example, if a key of
type string is not specified, its default value is "", and a key of type uint32 would be 0.