BUG#: 1911 TITLE: SampleProvider breaks build on Linux
DESCRIPTION: Took a blind whack at fixing this while at home on a Windows
system by modelling the creation of the CIMObjectPath after existing bits of code.
With any luck, this *may* prevent some build breaks before I have a chance to
build on Linux first thing in the morning.