BUG#: 1524 TITLE: Memory leaks in CIMxmlIndicationHandler
DESCRIPTION: Added the suggested workaround. With this change Purify does not show the memory leak, but looking at the code and the trace does not convince that the changes have addressed the root cause for the leak. For a later release the root cause of the memory leak may need to be identified and appropriate fix needs to be made.