BUG#: 1912 TITLE: Improved detection of GCC and version for temp object creation.
DESCRIPTION: Previously, determination of the use of GCC was done by specific platform or OS (which had the potential to be wrong in some cases).  In addition, the version of GCC was ignored when deciding if temporary objects had to be generated or not.  The determination of the use of GCC is now always correct and the version of GCC is now taken into account when making the decision.