BUG#: 1625 TITLE: Fix providermanager shared library loading on linux
DESCRIPTION: Ensure that the PEGASUS_DEST_LIB_DIR variable is set to the directory that holds Pegasus shared libraries.  Use this value to locate ProviderManager and IndicationHandler libraries when dynamically loading them.  A FileSystem::buildLibraryFileName method is added to consolidate handling of platform differences in library names and to avoid replicating this logic yet again.