BUG#: 1462 TITLE: Hostname validity checks applying out of date conventions
DESCRIPTION:
Currently, there are some areas of the Pegasus codebase applying an out of date
convention for distinguishing hostnames from an IP address in dotted quad form.
It is based on more strict former requirements for hostnames.

Those rules guaranteed that host and domain elements must start with a letter,
making isalpha/isdigit on the first character a reasonable quick-and-dirty way
to guess if a particular string contained a hostname or IP address.  But that
convention was superceded later by RFC 1123 (link below) to allow digits 0-9 to
begin a hostname as well, making the check a bit trickier to do correctly.

From RFC 1123:
 - The syntax of a legal Internet host name was specified in
   RFC-952 [DNS:4]. One aspect of host name syntax is hereby
   changed: the restriction on the first character is relaxed to
   allow either a letter or a digit.
 - If a dotted-decimal number can be entered without identifying
   delimiters, then a full syntactic check must be made, because
   a segment of a host domain name is now allowed to begin with a
   digit and could legally be entirely numeric (see Section 6.1.2.4).
   However, a valid host name can never have the dotted-decimal form
   #.#.#.#, since at least the highest-level component label will be
   alphabetic.

The algorithm used has been updated accordingly.