BUG#: 1864 TITLE: Indication Service must validate key property values
DESCRIPTION: Modified the Indication Service to handle the CreationClassName key property in the same way it handles the SystemCreationClassName and SystemName key properties (i.e. when an Indication Filter or Handler instance is created, if the key property does not exist or has a NULL value, the Indication Service adds the property with the correct value or sets the property to the correct value).  Modified the Indication Service to validate the value if the property does exist and has a non-NULL value.