BUG#: 2025 TITLE: CommMechanism attribute not included in template
DESCRIPTION: Corrected.  This attribute now shows up in the output.  Note
that there is an open issue of whether the value in this attribute should be
the property value (i.e. 2) or the the corresponding value qualifier value.
The SLP Template specification is ambigious on this issue.  Bug 2042 discusses
this issue.