BUG#: 2097 TITLE: benchmarkProvider returns invalid Property name
DESCRIPTION: The benchmarkProvider was generating property names from ...0000 to ...<n-1> instead of from ...0001 to ...<n> as defined in the mof.  In addition, (1) corrected serveral class definitions in the mof, (2) added a test to the Client to be sure that the correct number of Properties are returned on enumerateInstances and and getInstance and (3) changed value of LocalOnly to false.