BUG#: 2042 TITLE: SLP Provider. Use of enumerated values vs. value of value qualifier in attributes
DESCRIPTION: Corrected to use the value in the value qualifier instead of the
property value.  Thus, for the communications type it outputs cim-xml instead of
2.  Note that this is also a problem in the slp template specification since the
values are correctly defined in the enum attribute template but the spec description
clearly states that the value must be acquired, not the value from the value qualifier.