BUG#: 2132 TITLE:CMPI/FilesAndDirectories/tests Failure
DESCRIPTION:
Problem was that the ObjectPath was set by the CIMOM to have the result of
gethostbyname, which is just fine, except in this case. So the provider now
sets the ObjectPath hostname with the result of CSName (which is localhost if
compiled with SIMULATED flag). All of this new code is between SIMULATED ifdef