BUG#: 2044 TITLE: Interop test fails on windows and Linux
DESCRIPTION: Problem was with schema make.  However, since the test
proceeded to an internal step before the problem was found, added tests
for existence of the required classes as part of the start of interop
poststarttests.  now will fail immmediatly with error message about missing
classes if classes not in repository.