BUG#: 2044 PEP#: DELETE ONE OF THESE LINES TITLE: Interop test fails in Linux and windows
DESCRIPTION:   Updates makefile for interop schema to compile the CIM and PG interop classes both when the PEGASUS_ENABLE_SLP and the PEGASUS_HAS_PERFINST variables are set.  the problem was caused  because the conditional compiles were set to compile the interop provider
when PEGASUS_HAS_PERFINST was set but not the classes required by this provider.  Note that another commit added tests to the provider test
so that any class errors would be caught at start of test.