BUG#: 2163 TITLE: Indication not delivered because no check for derived class match
DESCRIPTION: Modified the IndicationService to use the existing Subscription Classes hash table (which was designed for fast lookup of subscriptions that match a specified indication class/subclass and source namespace) to retrieve the subscriptions that match the generated indication, then verify that each subscription included by the provider in the SubscriptionInstanceNamesContainer is also in the list of matching subscriptions retrieved from the Subscription Classes hash table.  Also, added a test case that generates an indication of a subclass.