BUG#:2230 TITLE: TestClient Failure running Nightly Tests DESCRIPTION:
This a day one bug that got exacerbated by PEP140/phase2 changes.
The hang was because not all the message responses got to the
resequencer in Server/CIMOperationRequestDispatcher.cpp , in
which case it never sent out the completion flag to the
client indicating the chunks are complete (hence the client
waits until timeout).

The culprit is MessageQueueService::_req_proc in
MessageQueueService.cpp. It does not loop on all queued
AsyncOpNodes queued by cimom::_routing_proc. It is possible that
many messages may get queued before the semaphore signal can
be processed (which is _req_proc). Objects like
CIM_ManagedSystemElement can cause many threads and evokes
the problem sporadically (CPU usage and number of CPUs play
a role in whether it failed or not)