BUG#: 2241 TITLE:  TestClient generates xml validation error with associations
DESCRIPTION:  Problem in SampleFamilyProvider where there was a path that could return responses to enumerates/associators calls with
the path component not set.  Removed this extra path since it was unnecessary in any case.  This did not fail before because the errors were
discarded as part of the aggregation so this was never noticed.