BUG#: 2276 TITLE: Removed unneeded code in #ifdef.
DESCRIPTION: There was some code that was an #ifdef that references a field called sa_restorers.  After looking at Linux, Solaris, HP-UX and AIX header files it seems this field is only defined on Linux and only when used in kernel mode.  Since Pegasus should never be used on Linux in kernel mode I went ahead and removed the #ifdef'd code.