PEP#: 187 TITLE: PEP:187 - SSL Certificate Management Enhancements - Phase 1.
DESCRIPTION: This submission includes 'ssltrustmgr' CLI that can be used by system admiistrators to manage the CRL store and the Trust stores on a local system. This CLI will not function fully until the certificate management provider code is submitted. The provider code and the CLI tests are planned for Phase 2 submission. Code changes are tested on HP-UX and Windows.