BUG#: 2609 TITLE: Change PERFINST statistics compile option from normally off to normally on
DESCRIPTION: This changes the env variable used to control this compile option and
changes it from normally off tonormally on.  The old variable was PEGASUS_HAS_PERFINST.
The new one is PEGASUS_DISABLE_PERFINST.  If you do not set the PEGASUS_DISABLE_PERFINST
the statistics will be enabled.   Note there was a special flag in PLATFOMR_OS400 to
disable statistics and we left that for the moment.