BUG#: 2472 TITLE: CLI getinstnace interpreted as delete instance
DESCRIPTION: Corrected.  This was a logic error in the inclusion of an
interactive function that allowed doing get instance interactively so that
if you only supply the class name on a get instance it returns a list of the
existing instances using enumerateinstance and offers to display one of them.
The same logic was put into delete instance and there was an error that caused
it to look like it was going to delete instead of display.