BUG#: 2565 TITLE: WQL does implement ISA
DESCRIPTION: Fixes the parser to accept the ISA syntax.  This is strictly
the parser fixes.  It does not add the code to process the ISA classname
result.  That will be added later.  This, however, should pass the SNIA
CTP tests for this function the way most people use that test.