PEP#: 187 TITLE: SSL Certificate Management Enhancements
DESCRIPTION: Added functionality to update trust store cache on deletion of a certificate from either the trust store or the CRL store. Created a new SSLContextManager class that owns the SSLContext objects. CIMServer and Certificate providers use SSLContextManager to create a new context and to update the trust store cache.