PEP#: 208 TITLE: PEP208 Repository Upgrade utility related files.
DESCRIPTION: Today, when a user wants to install a new version of CIM schema into a namespace, the existing namespace must be deleted and a new namespace created. Any information created by the users in the old namespace such as classes and/or instances of Provider registrations, Namespace authorizations, Indication subscriptions etc will be lost and the user would have to recreate them. Added a utility to support repository upgrade functionality. Added a sample Special Processing Module.