BUG#: 2731 TITLE: Correct Compile Failure buiding WQLYACC.cpp on HP-UX
DESCRIPTION: the yacc output was putting in a pragma with
the file name changed from uppercase to lower case. This was
corrected by using a different version of bison/flex.  Note that
there were a number of other differences that showed up when
this was compiled with an up-to-date version of the utilities.
The resulting files pass the included parser tests.