BUG#: 2792 TITLE: Statistics provider does not catch datetime exception
DESCRIPTION: In the transform from microsecs to datetime, the
datetime creation does not include a try block and it gets datetime
exceptions, causing the provider to return exception.  Put try block
around this for time being and set time to zero. Note that this is a work
around patch so we can do further testing.  It puts out a message to
console and SHOULD NOT be considered a permanent patch.