PEP#:195 TITLE: CIM listener enhancements
DESCRIPTION

Adds a standalone CIM listener that works independent of the CIMOM. This is the first phase of check-in; there is still some tweaking and unit-test that needs to be done.  The correct function of the listener is dependent on "Bugzilla 1612 - sem_wait under Linux us now POSIX compliant."  I believe that Konrad will soon be checking in the implementation for time_wait, which this function depends on.