BUG#: 2810 TITLE: Add options to activate and deactivate statistics gathering.
DESCRIPTION: Adds simple command line options (on, off, status) to change
and view the statistics gathering flag in the cimObjectManager.  Note that
the bug proposes a longer term solution but this serves as the basis for testing.
The function is turned on with testinterop on and turned off with testinterop off.
Note that the code was refactored in this test program to make the set of fuctions
it performs more logical and to add tests so that a diff will show a large number
of changes.