BUG#: 2651 TITLE: Stopwatch does not allow accumulation
DESCRIPTION: Add start() and stop() methods to Stopwatch to 1) allow the
measurement of specific time intervals, and 2) allow accumulation of multiple
intervals. That is, the Stopwatch can be started and stopped multiple times and
the sum of multiple intervals will be returned when getElapsed is called. Update
numerous test cases that use this object.