BUG#: 3114 TITLE: Fixed stack corruption problem.
DESCRIPTION: A very common problem: the third parameter to StringFromGUID2() is the number of CHARACTERS in the string, not the number of BYTES.  Since this is a UCS-2 string, a simple sizeof the variable is not enough, it must also include the size of the characters in the calculation. I moved such a calculation to a place a little earlier in the code and then used the calculated value.