BUG#:3076 TITLE: connectLocal() auth fails on Windows w/domain
DESCRIPTION:

Put in proposed fix, but commented it out, since the secur32.lib
does not seem to come std.  I had to pull it in from the SDK.  But this
is really the only right way to fix this.  The other ways are more complicated
and do not ensure we are getting the username of the thread.

Alex, if you grep for "3076" and uncomment those lines, you should be
able to try it.  I tested it against a DC and the local system.

I will put this on Tues call.