BUG#: 3207 TITLE: Fixed program popup error.
DESCRIPTION: Was getting a popup (on Windows) saying that the application terminated abnormally.  That was caused by an exception being propagated from main().  That occurred because a connect() call in main() was not done inside a try/catch.  So, I moved the connect() call into an existing try/catch.

Also, made some existing catch by references catch by const references.