BUG#: 3248 PEP#: TITLE:Compile failure building WQLYACC.cpp
DESCRIPTION: This proably derived from compiling the grammar change from
bug 3084 with the windows bison. Although the change was tested on linux
and windows, it appears that this was not compatible with HP compiler.
Recompiled with current linux bison and commited the update files.  There
was NO change to WQL.l or WQL.y