BUG#: 3390 TITLE: Use wrappers for conversion between Java and C++ objects in JMPI
DESCRIPTION:
I have rewritten the JMPI code to use a macro/template combination when it
converts between Java and C++ objects.  This was initially done to validate that
incoming java pointers point to actual C++ objects.

In the future, this could be used to detect memory leaks in the JMPI code.