BUG#: 3430 TITLE: Added getPortNumber() method.
DESCRIPTION: If the creator of the object specifies a port of 0 (indicating that the system should determine the port number to use) then the actual port number is queried and will be returned.  Previously, in this case 0 was returned.

Also took the opportunity to change another method to be const() since it didn't (and shouldn't have) changed the associated object.