BUG#: 3488 TITLE: Fixed buffer overflow problem which could cause crash.
DESCRIPTION: For some reason, the program name was copied from argv[0] and the buffer it was copied to was only 40 characters long.  The full path was specified in argv[0] and on my system that was around 60 characters, therefore other data was being overwritten.

I changed it so that a pointer to argv[0] is kept instead of copying the data.