BUG#: 3456 TITLE: CIM Server Crashes - No  threadsafe code in FamilyProvider
DESCRIPTION: The _Hostname object was not threadsafe.  This was recoded
to use the stack and not create an object for this information.  Note that
we did this generally so that the complete input path is made available
also eliminating the passage of namesapce as a separate object. Test on windows
with local and 10 hour test showing no memory loss and no failures.