BUG#: 3578 TITLE: Removed unnecessary include files.
DESCRIPTION: Also fixed a serious, but very unlikely to occur bug where a delete is followed by a new.  If the new fails then the value in the pointer that was just deleted remains.  Thus, making the likelyhood of a double delete very high.  Added an assignment of NULL to the pointer after the delete to prevent such an occurrence.