BUG#: 3325 TITLE: 'Fixed' problem with compiling C files.
DESCRIPTION: This is a hack, but it at least allows C files to compile properly.  It puts a C++ specific option as a part of the C++ compiler line so that it doesn't get used by the C compiler - as it is not a valid option to the C compiler.