BUG#: 2653 TITLE: SMI-S scoped property not implemented in Pegasus WQL
DESCRIPTION: Implemented in the same manner as the WQL ISA, with a
flag that allows it only to be compiled if the compile flag
PEGASUS_SNIA_EXTENSIONS is set.  This means that nobody but SNIA
users need have these extensions enabled.  Note that these are
enabled only in the parser, there is no evaluation support, etc.
Also the ParserCLI was expanded to provide negative tests so that
the non-existence of these functions in normal compile mode is also
tested.