BUG#: 2185 TITLE: Property poorly defined in PG_CIMXMLCommMech mof
DESCRIPTION: New property added to class to do enum instead of string
to make this more precisel.  InteropProvider and slp provider modified
to use this property.  Interop still fills out the old namespacetype
property but slp uses namespaceAccessProtocol.

Note that this also corrected an error in the access for PG_CIMXML..
from slp in that it was getting with deepinheritance = false and that
must be true.