BUG#:2939 TITLE:DynListener is in Pegasus directory
Description:

Somehow, I forgot to take out the platform dependencies surrounding
the DynListener/Service being built.  The platform dependencies were taken out
of the pegdynlistener library and the TestDynListener binary a while ago.

This is my oversight.  If this causes an issue on a platform (I don't think it will), we can exclude the
platform in the Makefile for now.