BUG#: 3183 TITLE: Backwards-compatible handling of CIM_ERR_NOT_SUPPORTED
DESCRIPTION:

Modified the OperationAggregate::resequenceResponse() method to keep a count of the NOT_SUPPORTED exception responses received and to clear the exceptions except in the case of the final response. For the final response, it checks to see if all of the previous responses have been NOT_SUPPORTED exceptions, and if that's the case, then it allows the NOT_SUPPORTED response to continue on to the HTTP layer where a client will receive it. Otherwise, if any other error types or any data was received, then the last NOT_SUPPORTED response is cleared as well.  This behavior is consistent with 2.4.1.