BUG#: 3280 TITLE: Certificate based authentication should perform additional authentication checks
DESCRIPTION: Added a validateUser interface to the AuthenticationManager and Authenticator interfaces. Updated the Authentication handlers to implement the validateUser interface. Updated the PAM StandAlone Process to also support this interface. Updated the HTTPAuthenticatorDelegator to invoke the validateUser interface before setting the user name in the Identity Container during Certificate based authentication.