BUG#: 3935 TITLE: cimserver hang waiting for mutex
DESCRIPTION: This is a work-around for a C-Cover problem. There appears to be a problem associated with the use of the cov_write call in a multi-threaded environment.  Since we are now able to shutdown the cimserver successfully, the calls to cov_write are not longer necessary and can be removed.