BUG#: 3803 TITLE: Classinfo attribute in SLP Template
DESCRIPTION: Was acquiring the attribute as a string and should have been
uint16.  Modified.  Now produces either a) no attribute at all if there are
no ClassInfo properties in the instances or b) list of the property values
for each classInfo with 0 as the default if there are any valid classinfo
properties in the CIM_namespace instances. Note that there is still an issue
between the DMTF specs and the SNIA SMIs spec in that the dmtf spec defines
the property value and the SNIA defines the value qualifier as output. We
are following the DMTF spec at this point.