BUG#: 3940 TITLE: CLI does not allow domains to be specified in username field
DESCRIPTION: Removed the code that allowed substituting @ for " on the command line.
Note that this code is removed completely so that anybody using this substitution will
break as of this checkin.  In the future, any use of the quote withing quoted strings must
use the platform defined escape mechanism (typically the backslash)