PEP#: 244 TITLE: Server Profile Implementation
DESCRIPTION:

Initial checkin to non-release branch for testing the modifications required for native Server Profile support in Pegasus. It includes a significant amount of code cleanup (although I'm sure there is still more to do) as well as support for the classes in the Server Profile that were previously supported by creating static instances in the repository. There is still work to do in a number of areas and a few open issues, but this should serve as a reasonable base for completing a fully compliant SMI-S Server Profile implementation.