PEP#: 244 TITLE: Server Profile Implementation
DESCRIPTION:

Add test provider for Server Profile implementation. The test provider implements the pieces required from a vendor in order to pass the SNIA CTP test for the Server Profile, serves as an example for how to properly register an SMI-S provider to take advantage of the new Server Profile implementation, and can be used to test the Server Profile implementation through the automated nightly tests (updating the Interop provider test program is still needs to be done).