BUG#: 4693 TITLE: Undesirable message logged when an exception is thrown by a provider
DESCRIPTION:

Modified DefaultProviderManager.cpp to set the provider and server start times on the response message before the provider is invoked, then set the provider end time after exceptions have been caught.

Also modified the file so that the identical blocks of catches can be replaced by a single macro.