BUG#: 4788 TITLE: pegtest user should be converted to use PEGASUS_TEST_USER_ID
DESCRIPTION:
This changes the hardcoded refereence to the user pegtest to use the new
env var PEGASUS_TEST_USER_ID which defaults to guest. It also adds the
logic to configure the users, if not doing PAM authentication, prior to
running the tests. It also includes the logic to make it a standalone tests.
IE it can be run directly outside of the testsuite and it will set up its
dependencies. It, as all other single tests do, presumes that the
repositories are pre-built.