BUG#: 4788 TITLE: pegtest user should be converted to use PEGASUS_TEST_USER_ID
DESCRIPTION: reverting the part of the change to put pegtest back.
until it can be resolved what PEGASUS_TEST_USER_ID and PASS should default to,
guest, pegtest or something else.

The original change caused the nightly test to fail. When this was submitted
the assumption was that since the machines that used pegtest were also running
the standardtest using the guest userid that the guest user id was a vlaid user.
Apparently that assumption was wrong and guest is not a valid user id.