PEP#: 233 TITLE: Update EmbeddedInstance branch with fixes for CMPI
DESCRIPTION:

A number of issues have popped up in the past week or so regarding EmbeddedInstance and EmbeddedObject types when implementing providers via the CMPI interface. Some necessary checks were not in place for EmbeddedObject and EmbeddedInstance types, and there were also issues because CMPI does not distinguish between the two types, so additional code had to be added to correct any mismatches between properties being returned as EmbeddedObjects that should have been EmbeddedInstances.