BUG#: 5065 TITLE: Nightly Build failure in RHEL4 U2 in TestOOPModuleFailure test case
DESCRIPTION: Modified ProviderRegistrationManager::updateProviderModuleStatus to hold the provider registration table lock while getting the current OperationalStatus property value, before updating the OperationalStatus property, thus avoiding problems caused by using a stale value of the OperationalStatus.  The code that looks up the provider module in the provider registration table, and gets the Operational Status from the module instance in the table, was consolidated into a new private method ProviderRegistrationManager::_getProviderModuleStatus, since it is now needed by both ProviderRegistrationManager::updateProviderModuleStatus and ProviderRegistrationManager::getProviderModuleStatus.  The Stopping status was modified to not replace the existing status, but be appended to the status array, to qualify the status.  In that way, if the disable fails, the previous status is not lost.